Move nodes into a module
