+++ apps/pipelines/nodes/types.py
@@ -1,0 +1,6 @@
+from typing import TypeAlias
+
+LlmProviderId: TypeAlias = int
+LlmModel: TypeAlias = str
+LlmTemperature: TypeAlias = float
+PipelineJinjaTemplate: TypeAlias = str  # TODO: type PipelineJinjaTempate = str in python 3.12
